Revert "Ben/devel"

+++ nixos/modules/hardware/ksm.nix
@@ -26,13 +26,13 @@ in {
     systemd.services.enable-ksm = {
       description = "Enable Kernel Same-Page Merging";
       wantedBy = [ "multi-user.target" ];
-      script =
-        ''
+      after = [ "systemd-udev-settle.service" ];
+      script = ''
+        if [ -e /sys/kernel/mm/ksm ]; then
           echo 1 > /sys/kernel/mm/ksm/run
-        '' + optionalString (cfg.sleep != null)
-        ''
-          echo ${toString cfg.sleep} > /sys/kernel/mm/ksm/sleep_millisecs
-        '';
+          ${optionalString (cfg.sleep != null) ''echo ${toString cfg.sleep} > /sys/kernel/mm/ksm/sleep_millisecs''}
+        fi
+      '';
     };
   };
 }
